Finished playable game
In the game, the orange arrow tries to eat the red squares, but these make it lose health. Your goal is to make sure the arrow doesn't eat the red squares and instead takes as many green circles as possible. You can do this by adding and removing red squares. If you click on the map, it adds a new one, when you click on an existing one, it gets deleted.