Replaced lives with a health bar
-Replaced the lives with a health bar.
This allows the health to slowly deplete over time. This gives you a reason to go for the salads, even if you didn't get hit by a burger. It also adds a better way of making the game harder the longer you play.
This also allows the burgers and salads to do more or less than 1 live of damage/healing for every item picked up.

-Made the player speed increase over time.