Added better highlights to hamburger placement
Added a preview of where you'll place a new hamburger to make it clear you have to use the mouse to place them. Also made the hamburgers slightly transparant when you hover over them with your mouse.
You can now use the 'r' key to reset when it's game over. This is so you don't have to restart the game every time.
Hamburgers don't spawn close to the player anymore. Previously, a hamburger could just spawn next to the player and get eaten immediately, which felt very unfair.